feat(search): extract markdown text preview popover and add it to Platform search page
